04: add simple usage, change argument declaration

+++ testforfun/04/4-3.c
@@ -7,7 +7,8 @@
 int i,j;
 
 // usage: int check_argument(<number>)
-// use it to specific
+// use it to specific argument number is correct
+/* 若使用者在程式後加的參數並非四個，則程式終止 */
 int check_arguments(int argnum, int num)
 {
     if (argnum != (num+1) ) {
@@ -18,6 +19,8 @@ int check_arguments(int argnum, int num)
     }
 }
 
+// usage: int check_last_bit(<32/64>)
+// use it to check last bit of binary number is correct
 void check_last_bit(int bitnum)
 {
     if(bitnum & 1) {
@@ -27,8 +30,9 @@ void check_last_bit(int bitnum)
     }
 }
 
-
-void trans_bits_work(int a, int bitnum)
+// usage: int trans_bits_work(<variable>)
+// use it to trans input variable to bit pattern
+void trans_bits_work(long long int a, int bitnum)
 {
     for(i=0;i<bitnum;i++) {
         j = a >> (bitnum-1-i);
@@ -36,6 +40,8 @@ void trans_bits_work(int a, int bitnum)
     }
 }
 
+// usage: int trans_bit_to_int_float(<argument(argv[number])>)
+// use it to trans bit pattern to integer or float variable
 void trans_bit_to_int_float(char* argument)
 {
     int result_int = 0;
@@ -47,6 +53,8 @@ void trans_bit_to_int_float(char* argument)
     printf("float: %f\n", trans_bits_float);
 }
 
+// usage: int trans_bit_to_int_float(<argument(argv[number])>)
+// use it to trans bit pattern to integer or double variable
 void trans_bit_to_int_double(char* argument)
 {
     unsigned long int result_int = 0;
@@ -58,6 +66,8 @@ void trans_bit_to_int_double(char* argument)
     printf("double: %lf\n", trans_bits_double);
 }
 
+// usage: int trans_bit_to_int_float(<argument(argv[number])>)
+// use it to check last bit of binary number is correct
 void trans_int_to_bit(char* argument, int bits)
 {
     long long int input_int;
@@ -67,6 +77,8 @@ void trans_int_to_bit(char* argument, int bits)
     printf("\n");
 }
 
+// usage: int trans_bit_to_int_float(<argument(argv[number])>)
+// use it to check last bit of binary number is correct
 void trans_float_to_bit(char* argument)
 {
     float input_float;
@@ -78,6 +90,8 @@ void trans_float_to_bit(char* argument)
     printf("\n");
 }
 
+// usage: int trans_bit_to_int_float(<argument(argv[number])>)
+// use it to check last bit of binary number is correct
 void trans_double_to_bit(char* argument)
 {
     double input_double;
